\name{default.zero.subst}
\Rdversion{1.1}
\alias{default.zero.subst}
\docType{data}
\title{
Default zero significance replacement
}
\description{
Since for Fisher's formula, signifances of zero would cause a problem (ln(0) == Inf), all such p-values are replaced with this number.
}
\usage{default.zero.subst}
\format{
  The format is:
 num 1e-06
}
\author{
Karl Kugler, karl@eigenlab.net
}
\keyword{datasets}
