\name{MADAM-package}
\Rdversion{1.1}
\alias{MADAM-package}
\alias{MADAM}
\docType{package}
\title{
MADAM - Meta-Analytical Data Aggregation Methods Toolbox
}
\description{
The MADAM package was developed in order to help reseachers and scientist working with meta-analysis. MADAM provides a set of method for generating test data, performing meta-analysis and visualizing the results.
}
\details{
\tabular{ll}{
Package: \tab MADAM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-10-27\cr
License: \tab LGPL \cr
LazyLoad: \tab yes\cr
}
}
\author{
Karl Kugler, UMIT
Maintainer: Karl Kugler, UMIT <Karl.Kugler@umit.at> 
}
\keyword{package}

