\name{MADAM-package}
\Rdversion{1.1}
\alias{MADAM-package}
\alias{MADAM}
\docType{package}
\title{
MADAM - Meta-Analytical Data Aggregation Methods Toolbox
}
\description{
The MADAM package was developed in order to help reseachers and scientist working with meta-analysis. MADAM provides a set of method for generating test data, performing meta-analysis and visualizing the results.\cr
Some functions were taken from the GeneMeta and RankProd packages in order to implement parallel versions of the existing methods. Please refer to the original authors if you make use of these implementations.
}
\details{
\tabular{ll}{
Package: \tab MADAM\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2010-06-29\cr
License: \tab LGPL \cr
LazyLoad: \tab yes\cr
URL: \tab http://madam.r-forge.r-project.org \cr
}
}
\author{
Karl Kugler, Matthias Wieser
Maintainer: Karl Kugler <karl@eigenlab.net>
}
\keyword{package}

